+++ schedule/__init__.py
@@ -1,12 +1,14 @@
 from flask import Flask, render_template, flash, redirect, url_for
+from flask_login import LoginManager, login_user, logout_user, current_user
 from flask.views import MethodView
 from flask_migrate import Migrate
-from schedule.model import db
-from schedule.model import Role, Timeinterval
+
+
 from schedule.user.models import Person
 from schedule.user.views import blueprint as user_blueprint
-
-from flask_login import LoginManager, login_user, logout_user, current_user
+from schedule.calendar.models import db
+from schedule.calendar.models import Role, Dutytype, Timeinterval
+from schedule.calendar.views import blueprint as calendar_blueprint
 
 
 def create_app():
@@ -21,6 +23,7 @@ def create_app():
     login_manager.login_view = 'user.login'
 
     app.register_blueprint(user_blueprint)
+    app.register_blueprint(calendar_blueprint)
 
     @login_manager.user_loader
     def load_user(user_id):
@@ -28,21 +31,7 @@ def create_app():
 
     @app.shell_context_processor
     def make_shell_context():
-        return {'db':db, 'Person':Person, 'Role':Role}
-
-    @app.route('/')
-    def index():
-        title = "Рабочий календарь"
-        return render_template('index.html', title = title)
-
-    @app.route('/smeny')
-    def smeny():
-        if current_user.is_authenticated:
-            title = "Смены дежурств"
-            return render_template('smeny.html', title = title)
-        else:
-            flash('Log in for access', 'alert-info')
-            return redirect(url_for('user.login'))
+        return {'db':db, 'Person':Person, 'Role':Role, 'Dutytype':Dutytype, 'Timeinterval':Timeinterval}
 
     return app
