+++ schedule/user/models.py
@@ -1,5 +1,5 @@
-from schedule.model import db
-from flask_sqlalchemy import SQLAlchemy
+from schedule.calendar.models import db
+#from flask_sqlalchemy import SQLAlchemy
 from flask_login import UserMixin
 from werkzeug.security import generate_password_hash, check_password_hash
